Move crawl, scrape into own modules.
