+++ src/events/crawler/index.js
@@ -1,59 +1,12 @@
 const arc = require('@architect/functions')
-const getSource = require('@architect/shared/sources/_lib/get-source.js')
-const crawler = require('./crawl')
-const cache = require('./cache')
+const crawl = require('./_crawl.js')
 
+/**
+ * Crawl a source, report the results
+ */
 async function crawlSource (event) {
   try {
-    /**
-     * Load the requested source
-     */
-    const source = getSource(event)
-    const { scrapers, _sourceKey } = source
-
-    const timeLabel = `Crawl ${_sourceKey}`
-    console.time(timeLabel)
-
-    /**
-     * Select the current scraper from the source's available scrapers
-     */
-    // TODO actually calculate latest start date; this hack works for now
-    const scraper = scrapers[scrapers.length - 1]
-
-    /**
-     * Prepare all the 'get' results to be cached
-     */
-    const results = []
-    // TODO maybe want to make this Promise.all once things are stable
-    for (let crawl of scraper.crawl) {
-      let { type, url } = crawl
-      if (typeof url !== 'string') {
-        const result = await url(crawler.client)
-        Object.assign(crawl, result)
-      }
-
-      const response = await crawler(type, crawl)
-      const data = response.body
-      const result = {
-        // Caching metadata
-        _sourceKey,
-        _name: scraper.crawl.length > 1 ? crawl.name : 'default',
-        // Payload
-        data,
-        type
-      }
-
-      results.push(result)
-    }
-
-    if (results.length !== scraper.crawl.length) {
-      throw Error(`Failed to crawl all requested 'get' sources`)
-    }
-
-    /**
-     * Cache the results
-     */
-    await cache(results)
+    await crawl(event)
 
     // Alert status to a successful crawl
     await arc.events.publish({
@@ -64,8 +17,6 @@ async function crawlSource (event) {
         status: 'success'
       }
     })
-
-    console.timeEnd(timeLabel)
   }
   catch (err) {
     // Alert status to a crawl failure
@@ -77,8 +28,6 @@ async function crawlSource (event) {
         status: 'failed'
       }
     })
-
-    console.log('Crawler error', event)
     throw err
   }
 }
